requirements.txt
ollama
sentence-transformers
faiss-cpu